+++ api/lib/application/admin-members/index.js
@@ -2,6 +2,7 @@ import Joi from 'joi';
 
 import { securityPreHandlers } from '../../../src/shared/application/security-pre-handlers.js';
 import { identifiersType } from '../../../src/shared/domain/types/identifiers-type.js';
+import { adminMemberController as srcAdminMemberController } from '../../../src/team/application/admin-member/admin-member.controller.js';
 import { adminMemberController } from './admin-member-controller.js';
 
 const register = async function (server) {
@@ -67,7 +68,7 @@ const register = async function (server) {
             allowUnknown: true,
           },
         },
-        handler: adminMemberController.saveAdminMember,
+        handler: srcAdminMemberController.saveAdminMember,
         notes: [
           "- Cette route est restreinte aux utilisateurs ayant les droits d'accès\n" +
             '- Elle permet de donner un accès à Pix Admin à un nouveau membre\n' +
