+++ api/tests/unit/application/admin-members/admin-member-controller_test.js
@@ -108,32 +108,4 @@ describe('Unit | Controller | admin-member-controller', function () {
       expect(statusCode).to.equal(204);
     });
   });
-
-  describe('#saveAdminMember', function () {
-    it('should return the serialized admin member saved', async function () {
-      // given
-      const attributes = { email: 'green.bot@example.net', role: ROLES.SUPER_ADMIN };
-      const savedAdminMember = Symbol('saved admin member');
-      sinon.stub(usecases, 'saveAdminMember').withArgs(attributes).resolves(savedAdminMember);
-
-      const serializedAdminMember = Symbol('serialized admin member');
-
-      const adminMemberSerializerStub = { deserialize: sinon.stub(), serialize: sinon.stub() };
-      adminMemberSerializerStub.deserialize.returns(attributes);
-      adminMemberSerializerStub.serialize.withArgs(savedAdminMember).returns(serializedAdminMember);
-
-      const dependencies = { adminMemberSerializer: adminMemberSerializerStub };
-
-      // when
-      const { statusCode, source } = await adminMemberController.saveAdminMember(
-        { payload: { data: { attributes } } },
-        hFake,
-        dependencies,
-      );
-
-      // then
-      expect(source).to.deep.equal(serializedAdminMember);
-      expect(statusCode).to.equal(201);
-    });
-  });
 });
